Fix windows_nat_target::fake_create_process ptid
During development of this series, I managed to introduce a bug that
let to fake_create_process being called.  That then resulted in GDB
crashes later on, because fake_create_process added a thread with an
incorrect ptid for this target.  It is putting dwThreadId in the tid
field of the ptid instead of on the lwp field.  This is fixed by this
patch.

I do however wonder why nobody has seen it this long.  The
fake_create_process code was added for
https://sourceware.org/bugzilla/show_bug.cgi?id=8153 and Chris said
back then, in:
  https://sourceware.org/legacy-ml/gdb-patches/2003-12/msg00479.html

  "If the main thread has exited, there is never any create process
  notification when attaching to a process.  That confuses all sorts of
  things."

I don't see any missing CREATE_PROCESS_DEBUG_EVENT when I try that
scenario on Windows 10.  I would have to try on Windows XP (the oldest
version of Windows we claim to support), but I don't have such a
machine handy.  Do note however, that GDBserver has never gotten an
equivalent to fake_create_process.

Change-Id: Iaee5d2deaa57c501f7e6909f8ac242af9b183215